Test with Java 21
Java 21 was released Sep 19, 2023. We want to announce Java 21 support
in October and would like the most used plugins to be compiled and tested
with Java 21.

The acceptance test harness and plugin bill of materials tests are already
passing with Java 21. This is a further step to improve plugin readiness
for use with Java 21 and for development with Java 21.

The change intentionally tests only two Java configurations, Java 17
and Java 21 because I believe that the risk of a regression that only
affects Java 11 is low. We generate Java 11 byte code with the Java 17
and the Java 21 builds, so Java 11 byte code is already tested.